reorg to separate opcode args from opcodes
